freeze: fix for compiling with Python3
To get the output of cython_freeze to compile with both Python 2 and
Python 3, I copied the contents of Python-3.1.1/Modules/python.c, with
some slight modifications.  The main issue is that Python 3 uses
wchar_t, while Python 2 uses char.  It also appears that the
Py_FlushLine/PyErr_Clear is not needed in Python 3.

Related to #434: "main() function generated by --embed doesn't compile
in Py3"